+++ cyvcf2/helpers.h
@@ -3,4 +3,8 @@
 #include <htslib/khash.h>
 
 int as_gts(int *gts, int num_samples, int ploidy, int strict_gt, int HOM_ALT, int UNKNOWN);
+int gt_types_012_from_int8(const int8_t *data, int num_samples, int ploidy,
+                           int strict_gt, int HOM_ALT, int UNKNOWN,
+                           int32_t *gt_types, int32_t *gt_idxs,
+                           int32_t *gt_phased);
 int32_t* bcf_hdr_seqlen(const bcf_hdr_t *hdr, int32_t *nseq);